GHG-T-12: Fix texture wrap sampling bug.
As explained in the comment in planet.frag, this was caused by
interpolating the vertices closest to the start of the texuture that
stood right next to the vertices closest to the end of the texture (at
the wrapping point around the sphere). These two lines got interpolated
between 0 and 1, so they sampled the colors across the entire texture in
a tiny little gap.

The simplest solution may be to just double up the vertices at the wrap
point, so they draw a bit more but they can sample accurately. However,
this will require some performance testing to verify.